IMPALA-9985: Bump Centos8 to 8.2, update Maven location
Bootstrapping GCC (both 4.9.2 and 7.5) on Centos8 hits the GCC
fixincludes bug messing up statx.h. Avoid this by using Centos 8.2 as
the build platform as it hopefully contains the glibc fixes for this
problem.

Additionally change the Maven download location to the new Apache
official archive site for a stable and maintained repo.

Change-Id: Ib622f1e636e8b17b72c67b3d48964acea20db9ff